backend/pyproject.toml
[project]
name = "cart-calculator"
version = "0.1.0"
description = "Distributed checkout API — clean MVP exploration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.22",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.5",
    "authlib>=1.3",
    "argon2-cffi>=23.1",
    "slowapi>=0.1.9",
    "limits[redis]>=3.10",
    "redis>=5.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0",
    "greenlet>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "asgi-lifespan>=2.1",
    "ruff>=0.7",
    "mypy>=1.11",
    "fakeredis>=2.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "C4", "PT", "RUF", "ASYNC", "S", "N"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S107"]
"alembic/versions/**" = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
files = ["app"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["slowapi.*", "authlib.*", "fakeredis.*", "argon2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"
